refactor: split oversized test responsibilities
