+++ tests/cli/test_mcp_runtime_compatibility.py
@@ -1,0 +1,69 @@
+from __future__ import annotations
+
+import json
+import sys
+from pathlib import Path
+
+from codex_usage_tracker.dashboard.api import generate_dashboard
+from codex_usage_tracker.diagnostics.api import run_doctor
+from codex_usage_tracker.pricing.api import (
+    annotate_rows_with_efficiency,
+    load_pricing_config,
+)
+from codex_usage_tracker.store.api import query_most_expensive_calls, refresh_usage_index
+from tests.store_dashboard_helpers import _make_codex_home, _write_pricing
+
+
+def test_pricing_annotation_and_doctor_pass(tmp_path: Path) -> None:
+    codex_home = _make_codex_home(tmp_path)
+    db_path = tmp_path / "usage.sqlite3"
+    dashboard_path = tmp_path / "dashboard.html"
+    pricing_path = _write_pricing(tmp_path / "pricing.json")
+    refresh_usage_index(codex_home=codex_home, db_path=db_path)
+    generate_dashboard(db_path=db_path, output_path=dashboard_path, pricing_path=pricing_path)
+
+    rows = query_most_expensive_calls(db_path=db_path, limit=1)
+    annotated = annotate_rows_with_efficiency(
+        rows, pricing=load_pricing_config(tmp_path / "missing-pricing.json")
+    )
+    assert annotated[0]["estimated_cost_usd"] is None
+    annotated = annotate_rows_with_efficiency(rows, pricing=load_pricing_config(pricing_path))
+    assert annotated[0]["estimated_cost_usd"] > 0
+
+    repo_root = tmp_path / "repo"
+    (repo_root / ".codex-plugin").mkdir(parents=True)
+    (repo_root / ".codex-plugin" / "plugin.json").write_text("{}", encoding="utf-8")
+    (repo_root / ".mcp.json").write_text(
+        json.dumps(
+            {
+                "mcpServers": {
+                    "codex-usage-tracker": {
+                        "command": sys.executable,
+                        "args": ["-m", "codex_usage_tracker.mcp_server"],
+                        "env": {"PYTHONPATH": str(Path(__file__).resolve().parents[2] / "src")},
+                    }
+                }
+            }
+        ),
+        encoding="utf-8",
+    )
+    plugin_link = tmp_path / "plugins" / "codex-usage-tracker"
+    plugin_link.parent.mkdir()
+    plugin_link.symlink_to(repo_root, target_is_directory=True)
+    marketplace_path = tmp_path / "marketplace.json"
+    marketplace_path.write_text(
+        json.dumps({"plugins": [{"name": "codex-usage-tracker"}]}),
+        encoding="utf-8",
+    )
+
+    report = run_doctor(
+        codex_home=codex_home,
+        db_path=db_path,
+        dashboard_path=dashboard_path,
+        pricing_path=pricing_path,
+        plugin_link=plugin_link,
+        marketplace_path=marketplace_path,
+        repo_root=repo_root,
+    )
+
+    assert report["status"] == "pass"
